Add timeout checking with SIGALRM
When 3 seconds passed after starting a test, SIGALRM is sent to the 
handler. Handler raises the SIGTERM to terminate the process.

This is a nasty approach. The ideal one could be skip to the result 
printing section. However the handler is not in the main function scope 
so I need to find other to perform this.